requirements.txt
streamlit
tensorflow
gdown
numpy
pillow
